Switch to using a specific version as specified in the code
Here's an example of generated code for something that shouldn't be
generated.  This code is apparently not being committed as the version
that was last committed is in the 0.8 when the latest tags are 0.10.x.

Relying so heavily on this means that code that's being released is not
committed to the repository.  That causes builds to not be reproducible.