NEEDS TESTING: Possible fix for conda uploads
Currently this slurps everyting out of the build directory.  If an
existing package were to be left in there it would get pulled in by
accident.  Binstar checks this by only uploading packages that match the
provided name, but that causes failures if there's any other files in
the conda-bld directory.

This change tightens up the glob to only pull in changes that match the
package name.  I'm not sure if I'm accessing the data correctly as the
only documentation as to what `job_data` is that I can find is a test.
Assuming the test data is correct, this will work.

Note, this needs testing.  I'm working with @tpowellcio to get access to
a Windows box to verify this.