Split working dir into staging and working dir.
Working directory is where the executing process should `cd` into
before execution. For the docker worker, this is relative to the
container file system.

Staging directory is where temporary files should be stored while
setting up a build, on the host machine. For the regular worker, this is
the same as the working directory.